+++ packages/terra-progress-bar/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added `valueText` prop to provide custom status update of progress.
+  * Added `ariaLabel` prop that labels the progress bar for screen readers.
 
 ## 4.35.0 - (November 13, 2023)